refactor: 모달 구조 정리
