--- src/components/atoms/Modal/Button/Button.types.ts
@@ -1,8 +1,0 @@
-import { ButtonHTMLAttributes } from "react";
-
-export type ButtonVariant = "confirm" | "default";
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  children: React.ReactNode;
-  variant?: ButtonVariant;
-}
