--- src/components/atoms/Modal/Button/index.tsx
@@ -1,13 +1,0 @@
-import { ButtonProps } from "./Button.types";
-import styles from "./Button.module.scss";
-export default function Button({
-  children,
-  variant = "default",
-  ...rest
-}: ButtonProps) {
-  return (
-    <button className={`${styles.button} ${styles[variant]}`} {...rest}>
-      {children}
-    </button>
-  );
-}
