--- src/components/atoms/Modal/Typography/index.tsx
@@ -1,8 +1,0 @@
-import styles from "./Typography.module.scss";
-
-interface TypographyProps {
-  children: React.ReactNode;
-}
-export default function Typography({ children }: TypographyProps) {
-  return <div className={styles.title}>{children}</div>;
-}
